feat(app): implement basic game play
In this session I got the app to deal cards and lay them out on the screen. Still working on trying
to get further actions to be accepted. Namely, I'm stuck trying to get the "hit" button to work. The
method dies because it can't find the correct hand to associate with the game and player. Focus on
decks class `dealOneTo	method